+++ docs/.vuepress/changelog.env.js
@@ -1,3 +1,3 @@
 const SHOW_CHANGE_LOG = true;
-const SHOW_PUBLISH_CLI = false;
+const SHOW_PUBLISH_CLI = true;
 module.exports = { SHOW_CHANGE_LOG, SHOW_PUBLISH_CLI };