+++ packages/webpack-loader-publish/src/config.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_PUBLISH_OPTIONS = {
+  port: '3005',
+  method: 'POST',
+  path: '/web-publish/publish',
+} as const;