+++ packages/webpack-loader-publish/src/publish-inner.ts
@@ -1,10 +1,29 @@
 import { localPublish } from 't-comm';
-import { PUBLISH_BASE_OPTIONS } from './config-publish';
+import { DEFAULT_PUBLISH_OPTIONS } from './config';
 
 
-export function innerPublish(publishEnv: string) {
+export function innerPublish({
+  publishEnv,
+  host,
+  port = DEFAULT_PUBLISH_OPTIONS.port,
+  method = DEFAULT_PUBLISH_OPTIONS.method,
+  path = DEFAULT_PUBLISH_OPTIONS.path,
+}: {
+  publishEnv: string;
+  host: string;
+  port?: string;
+  method?: string;
+  path?: string;
+}) {
+  if (!host && publishEnv !== 'devcloud') {
+    console.log('请从环境变量注入 `host`，如 process.env.VUE_APP_LOCAL_PUBLISH_HOST');
+    return;
+  }
   localPublish({
-    ...PUBLISH_BASE_OPTIONS,
+    host,
+    port,
+    method,
+    path,
     publishEnv,
   } as any);
 }