+++ packages/webpack-loader-publish/CHANGELOG.md
@@ -1,3 +1,22 @@
+## <small>0.0.7 (2024-09-29)</small>
+
+* fix(publish): 修复devcloud时无需host ([9d83ce3](https://github.com/novlan1/plugin-light/commits/9d83ce3))
+
+
+
+## <small>0.0.6 (2024-09-29)</small>
+
+* feat(publish): 增加bin命令 ([0348326](https://github.com/novlan1/plugin-light/commits/0348326))
+* feat(webpack-loader-publish): 去除敏感信息 ([9ac333b](https://github.com/novlan1/plugin-light/commits/9ac333b))
+* chore: add meta in pkg.json ([3f6088c](https://github.com/novlan1/plugin-light/commits/3f6088c))
+* chore: add release script ([defe6af](https://github.com/novlan1/plugin-light/commits/defe6af))
+* chore: update name ([02b2100](https://github.com/novlan1/plugin-light/commits/02b2100))
+* chore: update plugin-light-shared ([08a9790](https://github.com/novlan1/plugin-light/commits/08a9790))
+* chore: update tsconfig.json ([7d20e75](https://github.com/novlan1/plugin-light/commits/7d20e75))
+* test: update packages tests ([4897b59](https://github.com/novlan1/plugin-light/commits/4897b59))
+
+
+
 ## <small>0.0.5 (2024-09-18)</small>
 
 * fix: 修复路径错误 ([902218c](https://github.com/novlan1/plugin-light/commits/902218c))